+++ CHANGELOG.md
@@ -15,12 +15,14 @@ the code was deployed.
 
 - Endpoint for firmware state machine alerts (CU-v9ae26).
 - Storing and checking firmware state machine heartbeat messages (CU-v9ae26).
+- Alert Type to the Dashboard.
+- Tracking of when sessions are first responded to (CU-hjwfx2).
 
 ### Changed
 
 - Decoupled state machine from Sessions (CU-v9ae26).
 - Refactored and merged radar state machines (CU-v9ae26).
-
+- Use an enum for Alert Type in the DB and from Alert Lib instead of strings.
 
 ## [3.4.0] - 2021-06-21